requirements.txt
fastapi==0.115.12
uvicorn==0.29.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain-core==0.3.47
langchain-gigachat==0.3.8
langsmith==0.3.18
langgraph==0.3.18
requests==2.32.3 